dr_stocks-backend-main/microservices/requirements.txt
finnhub-python==2.4.10
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.31
sendgrid==6.9.7
requests==2.27.1
mysql-connector-python==8.0.28
pika==1.2.0